Slight refactor of message creation Move getCreatedMessageData() from the message stores getters to the ChatMessageUtils and move call to WebAPI's create method into the message store.
