+++ examples/flux-chat/js/stores/MessageStore.js
@@ -13,6 +13,7 @@
 var ChatAppDispatcher = require('../dispatcher/ChatAppDispatcher');
 var ChatConstants = require('../constants/ChatConstants');
 var ChatMessageUtils = require('../utils/ChatMessageUtils');
+var ChatWebAPIUtils = require('../utils/ChatWebAPIUtils');
 var EventEmitter = require('events').EventEmitter;
 var ThreadStore = require('../stores/ThreadStore');
 var assign = require('object-assign');
@@ -53,7 +54,7 @@ var MessageStore = assign({}, EventEmitter.prototype, {
   addChangeListener: function(callback) {
     this.on(CHANGE_EVENT, callback);
   },
-  
+
   removeChangeListener: function(callback) {
     this.removeListener(CHANGE_EVENT, callback);
   },
@@ -89,18 +90,6 @@ var MessageStore = assign({}, EventEmitter.prototype, {
 
   getAllForCurrentThread: function() {
     return this.getAllForThread(ThreadStore.getCurrentID());
-  },
-
-  getCreatedMessageData: function(text) {
-    var timestamp = Date.now();
-    return {
-      id: 'm_' + timestamp,
-      threadID: ThreadStore.getCurrentID(),
-      authorName: 'Bill', // hard coded for the example
-      date: new Date(timestamp),
-      text: text,
-      isRead: true
-    };
   }
 
 });
@@ -117,8 +106,12 @@ MessageStore.dispatchToken = ChatAppDispatcher.register(function(payload) {
       break;
 
     case ActionTypes.CREATE_MESSAGE:
-      var message = MessageStore.getCreatedMessageData(action.text);
+      var message = ChatMessageUtils.getCreatedMessageData(
+        action.text,
+        ThreadStore.getCurrentID()
+      );
       _messages[message.id] = message;
+      ChatWebAPIUtils.createMessage(message);
       MessageStore.emitChange();
       break;
